src/requirements.txt
Django==3.1.7
djangorestframework
drf-yasg
